Fixed set budget
When you set budget you can see when you come back

+++ lib/data/category/repository/category_repository_impl.dart
@@ -25,7 +25,7 @@ class CategoryRepositoryImpl extends CategoryRepository {
     required String desc,
     required int icon,
     bool isPredefined = false,
-    int? budget = -1,
+    double? budget = -1,
   }) async {
     await dataSources.addCategory(Category(
       description: desc,
